+++ core.js
@@ -1,3 +1,5 @@
+import { cache, pressed } from './Cache.js';
+
 $(document).ready(function() {
   var toggle = false;
   $('.splash .settings').click(function() {
@@ -441,43 +443,8 @@ $(document).ready(function() {
   });
 
 
-  var cache = {
-    size: [],
-    start: [],
-    stop: [],
-    press: false,
-    zoom: 0,
-    zoomed: false,
-    pos: function(x, y) {
-      return [x-$('#editor')[0].getBoundingClientRect().left, y-$('#editor')[0].getBoundingClientRect().top];
 
-    },
-    svgID: 0,
-    set ele(val) {
-      this.svgID = val;
-    },
-    get ele() {
-      return $('#editor #'+this.svgID);
-    },
-    set hoverEle(val) {
-      this.hovering = val;
-    },
-    get hoverEle() {
-      return $('#'+this.hovering);
-    },
-    selectedElements: [],
-    dragTab: false,
-    mapKeysTo: 'selection',
-    moveAmount: 2
-  }
-  var pressed = {
-    handle: false,
-    shiftKey: false,
-    ctrlKey: false,
-    altKey: false,
-    cmdKey: false,
-    tabKey: false
-  }
+
   var tool = {
     toolType: 'line',
     name: false,
